+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 
 ## 0.21.0 (upcoming)
 
+* [EOS-5743] Modify step to retrieve admin and sis basepath from keos.yaml
 * [GS-3881] Adapt management-baas path according to keos version
 * [QATM-3773] Refactor creating cutom users/groups
 * [GS-3161] Add step to create custom JTW